Remove support for vim-dispatch
Dispatch is a great hack. It mostly works without problems if you just want
to, well, dispatch any programs.

But it has different adapters that all act a bit different, uses a blacklist
for certain escape sequences and it's not very configurable, which makes it
hard to include in other plugins.

Also because of Vim's textlock, you can't close the quickfix window that
dispatch opens, even if there are no matches at all. Consider this:

    function! s:foo()
      cclose
    endfunction

    autocmd FileType qf call s:foo()

    let &makeprg     = 'grep -Rn qu'.'ux .'
    let &errorformat = '%f:%l:%m'
    Make

This will close the quickfix window, but also jump to the first line of the
buffer and there's no way to avoid this.

If you need real async search without resorting to complicated hacks, use
Neovim instead.