Improved quickfix handling with title
This uses the new `title` argument with `setqflist`/`setloclist`, if
available to set the title with a suffix that allows to handle it via
a `FileType qf` autocommand later, too.

With this the qf windows will have mappings and title always (with
`-noopen` and after `:cclose/:copen`).

Fixes https://github.com/mhinz/vim-grepper/issues/40.