Add lines immediately to qf/loc window
Prior to this change we wrote all lines to a buffer and only populated the
qf/loc window at the end.

The interface of :Grepper did NOT change. -open still opens when processing ends
etc.

You can see the difference if you use a slower grep tool (e.g. "grep") on a huge
directory and then :copen or :lopen to watch the window populating line-by-line.

Closes #94.