highlight: remove any flags from query
When using `:Grepper -tool rg` with a query like `-uu foo` (via the
prompt) to also search hidden files, the `-uu` should not be included in
the highlighted query.

This might get changed to be smarter, e.g. to only consider the first
word, e.g. with "-uu foo dir" only "foo" is the pattern, but it would
need to handle "-uu 'foo bar' dir" also.