prevent disappearing cursor
When no search results are found, s:finish_up() shows a message to that
effect:

    echo 'No matches found.'

Unfortunately, this causes the cursor to disappear. This only affects
the case where no matches are found -- the other use of `echo` is to
report on how many matches were found, but this coincides with moving
the cursor to the results list (quickfix/location-list), and there's no
disappearing cursor.

This "fix" feels more like a workaround, but after sprinkling numerous
redraws around the affected code, it's the only solution I could come up
with, and seems pretty benign.

To repro the bug this addresses:

    git clone git@github.com:mhinz/vim-grepper.git
    cd vim-grepper
    git checkout 635c14284aea79117f2cc9d07c18c8f6da23d85b
    vim -u NONE --clean -c 'source plugin/grepper.vim'

Then run:

    :Grepper

and search for something that doesn't exist:

    > somequerywithnoresults