Make -query not imply -noprompt (#182)
Honestly I can't remember the rationale behind the original decision, but now it
is possible to pre-populate the query _without_ searching right away.

This is especially useful if you want to build your own commands. E.g. a query
for Markdown could be pre-populated with a regexp that matches only headers.