Fix regression around escape sequences
We mapped esc to cancel the prompt, but when using Vim (not an issue in Nvim due
to its proper key parsing via libtermkey), even arrow keys would cancel the
prompt because they emit sequences like \eOA, \eOB, etc.

Now we reverse the logic. The default mode is to cancel the prompt unless <cr>
(or the equivalent <c-m>) is typed.

This was once fixed in
https://github.com/mhinz/vim-grepper/pull/106/commits/2271e35f79a62b8cc8a28f5c91d03b09a3ceb722
but regressed when we removed the need for magic strings.

Fixes https://github.com/mhinz/vim-grepper/issues/221