Introduce GrepperQuery highlight group
The behavior changed in this commit: `8b78347`. The decision was to
unconditionally highlight the query like the user's colorscheme
highlights string literals.

I'd like to be able to customize this without changing how all string
literals look, so I've introduced a new `GrepperQuery` highlight group
that links to `String` by default. This means that there should be no
change by default.

I have not written tests for this but I have tested it in my own
configuration files by adding this line:

```vim
hi link GrepperQuery Normal
```

As seen in the diff, there is prior art for introducing new highlighting
groups in grepper already--`GrepperPrompt` is a highlight group that
controls the color of everything before the `>` (which displays the
search program and flags). There do not appear to be any tests for that
feature either, otherwise I would have cargo culted some to test this
feature.