Remove file/path completion for Grepper command
As far as I can tell, this does not make sense for the Grepper command.
The command opens the prompt, from which path completion is performed.

I suspect the right thing to do here is complete with all flags, but it
seems sensible to leave that for another commit to make changes clear.